src/main/java/com/codeonblue/controller/HomeController.java
package com.codeonblue.controller;

import com.codeonblue.service.ProductService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;

@Controller
public class HomeController {

    @Autowired
    private ProductService productService;

    @RequestMapping("/")
    public String home(){
        return "home";
    }

    @RequestMapping("/productList")
    public String getProducts(Model model){
        model.addAttribute("productList", productService.findAll());
        return "productList";
    }


}
